package for 0.9

+++ package/index.js
@@ -139,7 +139,7 @@ export default class SvelteCMS {
                 return [id, this.fields[id] || id];
             }))
         }, this);
-        let adminStore = conf.adminStore || conf.configPath || 'src/sveltecms.config.json';
+        let adminStore = conf.adminStore || conf.configPath || 'src/lib/sveltecms.config.json';
         if (typeof adminStore === 'string' && !this.contentStores[adminStore]) {
             let contentDirectory = adminStore.replace(/\/[^\/]+$/, '');
             let fileExtension = adminStore.replace(/.+[\.]/, '');
@@ -148,7 +148,6 @@ export default class SvelteCMS {
             adminStore = {
                 type: 'staticFiles',
                 options: {
-                    prependContentTypeIdAs: '',
                     contentDirectory,
                     fileExtension,
                 }
@@ -511,7 +510,7 @@ export default class SvelteCMS {
             // Run contentPreWrite hooks, and bail if there is an error
             try {
                 if (!options.skipHooks)
-                    await this.runHook('contentPreSave', items[i], this, { ...db.options, ...options });
+                    await this.runHook('contentPreSave', items[i], contentType, this, { ...db.options, ...options });
             }
             catch (e) {
                 e.message = `Error saving content ${items[i]._type}/${items[i]._slug}:\n${e.message}`;
@@ -530,7 +529,8 @@ export default class SvelteCMS {
                 items[i]._errors = e.message.split('\r');
             }
         }
-        await this.indexer.saveContent(contentType, items.map(i => this.getIndexItem(i)));
+        if (!options.skipIndex)
+            await this.indexer.saveContent(contentType, items.map(i => this.getIndexItem(i)));
         return Array.isArray(content) ? items : items[0];
     }
     async deleteContent(contentType, content, options = {}) {
@@ -544,7 +544,7 @@ export default class SvelteCMS {
             // Run contentPreWrite hooks, and bail if there is an error
             try {
                 if (!options.skipHooks)
-                    await this.runHook('contentPreDelete', items[i], this, { ...db.options, ...options });
+                    await this.runHook('contentPreDelete', items[i], contentType, this, { ...db.options, ...options });
             }
             catch (e) {
                 e.message = `Error deleting content ${items[i]._type}/${items[i]._slug}:\n${e.message}`;
@@ -560,7 +560,8 @@ export default class SvelteCMS {
                 items[i]._errors = e.message.split('\r');
             }
         }
-        await this.indexer.deleteContent(contentType, items.map(i => this.getIndexItem(i)));
+        if (!options.skipIndex)
+            await this.indexer.deleteContent(contentType, items.map(i => this.getIndexItem(i)));
         return Array.isArray(content) ? items : items[0];
     }
     newContent(contentTypeID, values = {}) {
